refactor: 공용 여부에 따른 문장 세트 로직 분리

+++ src/main/java/com/example/integration/service/PublicSentenceSetService.java
@@ -1,20 +1,19 @@
 package com.example.integration.service;
 
 import com.example.api.response.CustomException;
-import com.example.integration.config.event.SentenceSetViewedEvent;
 import com.example.integration.config.exception.ErrorCode;
 import com.example.integration.config.util.SecurityUtil;
+import com.example.integration.entity.RoleType;
 import com.example.integration.entity.Sentence;
 import com.example.integration.entity.SentenceSet;
 import com.example.integration.entity.User;
 import com.example.integration.entity.dto.sentence.PagingResponseDto;
-import com.example.integration.entity.dto.sentence.SentenceResponseDto;
+import com.example.integration.entity.dto.sentence.PublicSentenceResponseDto;
 import com.example.integration.entity.dto.sentenceSet.*;
 import com.example.integration.repository.SentenceRepository;
 import com.example.integration.repository.SentenceSetRepository;
 import com.example.integration.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -25,12 +24,11 @@ import java.util.List;
 
 @RequiredArgsConstructor
 @Service
-public class SentenceSetService {
+public class PublicSentenceSetService {
 
     private final SentenceSetRepository sentenceSetRepository;
     private final SentenceRepository sentenceRepository;
     private final UserRepository userRepository;
-    private final ApplicationEventPublisher eventPublisher;
 
     /**
      * 현재 사용자 조회
@@ -53,73 +51,92 @@ public class SentenceSetService {
     }
 
     /**
-     * 개인 문장 세트 조회
+     * 현재 로그인한 사용자가 관리자인지 확인
+     * @param user
+     */
+    private void checkAdmin(User user) {
+        if (!user.getRoleType().equals(RoleType.ADMIN)) {
+            throw new CustomException(ErrorCode.FORBIDDEN);
+        }
+    }
+
+    /**
+     * 공용 문장 세트 목록 조회
      * @param offset
      * @param limit
      * @return
      */
     @Transactional(readOnly = true)
-    public UserSentenceSetResponseDto getSentenceSetByUser(int offset, int limit) {
-        User user = currentUser();
+    public ListSentenceSetResponseDto getPublicSentenceSetList(int offset, int limit) {
         Pageable pageable = PageRequest.of(offset / limit, limit);
-        List<SentenceSet> sentenceSetList = sentenceSetRepository.findAllByUserId(user.getId(), pageable);
+        List<SentenceSet> sentenceSetList = sentenceSetRepository.findAllWhichPublic(pageable);
 
-        return new UserSentenceSetResponseDto(user, sentenceSetList);
+        return new ListSentenceSetResponseDto(sentenceSetList);
     }
 
     /**
-     * 특정 개인 문장 세트와 포함된 문장 목록 조회
+     * 키워드(title)를 바탕으로 공용 문장 세트 목록 조회
+     * @param keyword
+     * @param offset
+     * @param limit
+     * @return
+     */
+    @Transactional(readOnly = true)
+    public ListSentenceSetResponseDto searchPublicSentenceSetList(String keyword, int offset, int limit) {
+        Pageable pageable = PageRequest.of(offset / limit, limit);
+        List<SentenceSet> sentenceSetList = sentenceSetRepository.findAllWithKeywordWhichPublic(keyword, pageable);
+
+        return new ListSentenceSetResponseDto(sentenceSetList);
+    }
+
+    /**
+     * 특정 공용 문장 세트와 포함된 문장 목록 조회
      * @param sentenceSetId
      * @param page
      * @return
      */
     @Transactional(readOnly = true)
-    public SentenceSetAndPagingResponseDto getSentenceSetWithSentences(Long sentenceSetId, int page) {
+    public PublicSentenceSetAndPagingResponseDto getPublicSentenceSetWithSentences(Long sentenceSetId, int page) {
         SentenceSet sentenceSet = findSentenceSetWithId(sentenceSetId);
 
         // 개인 문장 세트 조회일 때
         if (!sentenceSet.isPublic()) {
-            User user = currentUser();
-
-            if (user.getId() == sentenceSet.getUser().getId()) {
-                // 문장 조회 후 최근 조회 이벤트 발생
-                eventPublisher.publishEvent(new SentenceSetViewedEvent(sentenceSetId));
-            } else {
-                throw new CustomException(ErrorCode.PRIVATE_SENTENCE_SET);
-            }
+            throw new CustomException(ErrorCode.PRIVATE_SENTENCE_SET);
         }
 
         // 10개 페이징
         Pageable pageable = PageRequest.of(page, 10);
         Page<Sentence> sentencesPage = sentenceRepository.findBySentenceSetId(sentenceSet.getId(), pageable);
 
-        SentenceSetResponseDto sentenceSetResponseDto = new SentenceSetResponseDto(sentenceSet);
+        PublicSentenceSetResponseDto sentenceSetResponseDto = new PublicSentenceSetResponseDto(sentenceSet);
 
-        Page<SentenceResponseDto> sentenceResponseDtoPage = sentencesPage.map(SentenceResponseDto::new);
+        Page<PublicSentenceResponseDto> sentenceResponseDtoPage = sentencesPage.map(PublicSentenceResponseDto::new);
 
-        PagingResponseDto<SentenceResponseDto> pagingResponseDto = PagingResponseDto.of(sentenceResponseDtoPage);
+        PagingResponseDto<PublicSentenceResponseDto> pagingResponseDto = PagingResponseDto.of(sentenceResponseDtoPage);
 
-        return new SentenceSetAndPagingResponseDto(sentenceSetResponseDto, pagingResponseDto);
+        return new PublicSentenceSetAndPagingResponseDto(sentenceSetResponseDto, pagingResponseDto);
     }
 
     /**
-     * 새로운 개인 문장 세트 생성
+     * 새로운 공용 문장 세트 생성
      * @param sentenceSetRequestDto
      * @return
      */
     @Transactional
-    public SentenceSetResponseDto createSentenceSet(SentenceSetRequestDto sentenceSetRequestDto) {
+    public PublicSentenceSetResponseDto createPublicSentenceSet(SentenceSetRequestDto sentenceSetRequestDto) {
+        // 사용자가 관리자인지 확인
         User user = currentUser();
+        checkAdmin(user);
 
         SentenceSet sentenceSet = SentenceSet.builder()
                 .title(sentenceSetRequestDto.title())
-                .isPublic(false)
+                .isPublic(true)
                 .user(user)
                 .build();
 
         SentenceSet savedSentenceSet = sentenceSetRepository.save(sentenceSet);
 
-        return new SentenceSetResponseDto(savedSentenceSet);
+        return new PublicSentenceSetResponseDto(savedSentenceSet);
     }
 
     /**
@@ -129,14 +146,22 @@ public class SentenceSetService {
      * @return
      */
     @Transactional
-    public SentenceSetResponseDto updateSentenceSet(Long sentenceSetId, SentenceSetRequestDto sentenceSetRequestDto) {
+    public PublicSentenceSetResponseDto updateSentenceSet(Long sentenceSetId, SentenceSetRequestDto sentenceSetRequestDto) {
+        User user = currentUser();
+        checkAdmin(user);
+
         SentenceSet sentenceSet = findSentenceSetWithId(sentenceSetId);
+
+        // 개인 문장 세트인 경우 수정 X
+        if (!sentenceSet.isPublic()) {
+            throw new CustomException(ErrorCode.FORBIDDEN);
+        }
 
         sentenceSet.updateSentenceSet(sentenceSetRequestDto.title());
 
         SentenceSet updatedSentenceSet = sentenceSetRepository.save(sentenceSet);
 
-        return new SentenceSetResponseDto(updatedSentenceSet);
+        return new PublicSentenceSetResponseDto(updatedSentenceSet);
     }
 
     /**
@@ -145,9 +170,16 @@ public class SentenceSetService {
      */
     @Transactional
     public void deleteSentenceSet(Long sentenceSetId) {
+        User user = currentUser();
+        checkAdmin(user);
+
         SentenceSet sentenceSet = findSentenceSetWithId(sentenceSetId);
+
+        // 개인 문장 세트인 경우 수정 X
+        if (!sentenceSet.isPublic()) {
+            throw new CustomException(ErrorCode.FORBIDDEN);
+        }
 
         sentenceSetRepository.delete(sentenceSet);
     }
-
 }
